+++ projects/ui/src/lib/components/file-upload/index.ts
@@ -1,0 +1,3 @@
+export { KuiFileUploadComponent } from './kui-file-upload.component';
+export type { KuiFileUploadMode, KuiFileUploadVariant } from './kui-file-upload.component';
+export type { KuiUploadFile, KuiUploadFileStatus } from './kui-upload-file.interface';